changed uint32_t for number of buckets to uint64_t to support values that do not fit in 32 bits

+++ include/clht.h
@@ -231,7 +231,7 @@ typedef struct ALIGNED(CACHE_LINE_SIZE) clht_hashtable_s
 inline uint64_t __ac_Jenkins_hash_64(uint64_t key);
 
 /* Hash a key for a particular hashtable. */
-uint32_t clht_hash(clht_hashtable_t* hashtable, clht_addr_t key );
+uint64_t clht_hash(clht_hashtable_t* hashtable, clht_addr_t key);
 
 
 static inline int
@@ -330,7 +330,7 @@ _mm_pause_rep(uint64_t w)
 /* ******************************************************************************** */
 
 /* Create a new hashtable. */
-clht_t* clht_create(uint32_t num_buckets);
+clht_t* clht_create(uint64_t num_buckets);
 /* initializes the necessary per-thread structures for the hash table */
 void clht_gc_thread_init(clht_t* hashtable, int id);
