+++ include/clht_lb.h
@@ -174,11 +174,11 @@ _mm_pause_rep(uint64_t w)
   *lock = 0;	  
 
 /* Create a new hashtable. */
-clht_hashtable_t* clht_hashtable_create(uint32_t num_buckets );
-clht_t* clht_create(uint32_t num_buckets);
+clht_hashtable_t* clht_hashtable_create(uint64_t num_buckets );
+clht_t* clht_create(uint64_t num_buckets);
 
 /* Hash a key for a particular hashtable. */
-uint32_t clht_hash(clht_hashtable_t* hashtable, clht_addr_t key );
+uint64_t clht_hash(clht_hashtable_t* hashtable, clht_addr_t key );
 
 /* Insert a key-value pair into a hashtable. */
 int clht_put(clht_t* h, clht_addr_t key, clht_val_t val);
